% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.meps.acs.health.literacy-package.R
\docType{package}
\name{dc.meps.acs.health.literacy-package}
\alias{dc.meps.acs.health.literacy}
\alias{dc.meps.acs.health.literacy-package}
\title{dc.meps.acs.health.literacy: Social Data Commons: Education & Training: Health Literacy Estimates}
\description{
Allows user to easily get health literacy estimates data from the SDAD Data Commons. Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Hanna Charankevich \email{hc2cc@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
